refactor: split DependencyAnalysisService into 3 SRP services + extract JSON parser
DependencyAnalysisService had grown to 610 LOC owning 4 responsibilities:
namespace dependency graphs, DI registration scanning, NuGet package listing,
and NuGet vulnerability scanning. The inline ParseNuGetVulnerabilityJson
method was 86 LOC with cyclomatic complexity ~27 due to a captured nested
local function (AddPackagesFromFramework).

Split into three single-responsibility services and extract the JSON parser
into a static helper. The old IDependencyAnalysisService interface and
DependencyAnalysisService class are deleted (no façade) per workflow.md's
"no backwards-compatibility shims" rule.

## New surface

- INamespaceDependencyService / NamespaceDependencyService — namespace graph
  + circular-dependency detection (DFS).
- IDiRegistrationService / DiRegistrationService — DI registration scan
  (AddSingleton/Scoped/Transient + keyed variants), service/impl type
  classification, typeof() argument resolution.
- INuGetDependencyService / NuGetDependencyService — package listing via
  evaluated MSBuild items + vulnerability scan via gated executor.
- NuGetVulnerabilityJsonParser (Helpers) — extracted parser. AddPackagesFromFramework
  promoted from a captured nested function to an explicit static helper.
  Cyclomatic complexity drops from ~27 to ~8.

## Updated consumers (all callers swap to the relevant new interface)

- AdvancedAnalysisTools: get_di_registrations / get_namespace_dependencies / get_nuget_dependencies
- SecurityTools: nuget_vulnerability_scan
- OrchestrationService: extract_and_wire_interface_preview (uses IDiRegistrationService)
- RoslynPrompts: analyze_dependencies (now takes both namespace + nuget services),
  guided_package_migration (INuGetDependencyService), security_review (INuGetDependencyService)
- ServiceCollectionExtensions: 1 registration -> 3 registrations
- Test infrastructure: TestBase + TestServiceContainer expose the 3 new services

## Backlog
- Closes `dependency-service-srp-split` (subsumes the inline-parser extraction).
