+++ ai_docs/backlog.md
@@ -2,7 +2,7 @@
 
 <!-- purpose: Open work only; contract for agents syncing backlog on ship. -->
 
-**updated_at:** 2026-04-08T15:50:00Z
+**updated_at:** 2026-04-08T16:30:00Z
 
 ## Agent contract
 
@@ -32,7 +32,6 @@ Ordered by severity: P2 (contract violations / real-world blockers) → P3 (refa
 |----|---------|------|-----|
 | `test-run-failure-envelope` | none | — | **P2 / reliability.** 2026-04-08 audits: `test_run` failed with MSB3027/3021 file locks when `testhost` still holds Roslyn DLLs (roslyn-backed-mcp 130615), and with a generic MCP bridge error `An error occurred invoking 'test_run'` with no structured payload (131317). Surface exit code, stdout/stderr excerpts, and whether the failure is retryable; document Windows concurrent-test behavior in the tool description. |
 | `audit-feasibility-companion-cli` | none | `workspace-payload-summary-modes` | **P2 / tooling.** A complete walkthrough of `ai_docs/prompts/deep-review-and-refactor.md` against a 34-project / 751-document solution does not fit in a single Claude Code conversation today, primarily because workspace-payload bloat tools (`workspace_load`, `workspace_status`, `workspace_list`, `roslyn://workspaces`, `get_nuget_dependencies`, `list_analyzers`, `get_msbuild_properties`, `find_unused_symbols`) exhaust the per-turn ~250 KB output budget by mid-prompt — observed against the 34-project ITChatBot.sln in the 2026-04-07 rw-lock audit. Consider shipping a `roslyn-mcp-audit` companion CLI that drives the deep-review prompt headlessly and emits the canonical raw audit file as an artifact, freeing the LLM from having to render every tool result back through context. The payload-summary-modes fix alone may lift this ceiling enough that a CLI is unnecessary — re-evaluate after that lands. |
-| `dependency-service-srp-split` | none | — | **P3 / refactor.** `DependencyAnalysisService` (now 610 LOC, up from 530) owns four responsibilities: namespace dependency graphs, DI registration scanning, NuGet package listing, and NuGet vulnerability scanning. Split into `NamespaceDependencyService`, `DiRegistrationService`, and `NuGetDependencyService`; `DependencyAnalysisService` becomes a façade or is deleted. Also covers `ParseNuGetVulnerabilityJson` (still inline, cyclomatic ~27, ~86 LOC with nested local function capturing outer variables) — extract `NuGetVulnerabilityJsonParser` static class and eliminate the nested function. Defer until a touch-the-class change is needed. |
 | `orchestration-migrate-package-extract` | none | — | **P3 / refactor.** `OrchestrationService.PreviewMigratePackageAsync` (still 102 LOC, cyclomatic 17) handles project-file and `Directory.Packages.props` mutation logic inline. Extract `BuildPackageReferenceEdit`, `BuildCentralVersionEdit`, and `BuildRemoveVersionAttributeEdit` helpers to make control flow readable and allow isolated testing of each mutation shape. Defer until a touch-the-method change is needed. |
 | `cohesion-metrics-source-gen-aware` | none | — | **P3 / accuracy.** `get_cohesion_metrics` mis-handles two common patterns: (a) `[LoggerMessage]` and `[GeneratedRegex]` source-generator partial methods are counted as separate clusters, so a class like `QuestionClassifier` with one real method (`Classify`) gets `Lcom4Score=4`, falsely implying SRP violation; (b) local helper methods (e.g., `CreateFailure`, `MapToQueryParameters` in `SysLogServerAdapter`/`GitRepositoryAdapter` per the 2026-04-07 ITChatBot rw-lock audit) are listed simultaneously as standalone methods AND inside the `SharedFields` array. Fix: filter out source-gen partials by syntax/attribute detection before clustering, and decide canonically whether a local helper appears as a method OR a shared field, not both. |
 | `flow-analysis-expression-bodied-lift` | none | — | **P3 / improvement.** `analyze_data_flow` and `analyze_control_flow` cannot inspect expression-bodied members (they fail with `InvalidOperation: No statements found in the line range...`). The error is actionable, but expression-bodied methods are common in modern .NET and developers often want flow analysis on complexity hotspots written in that style (concrete case from 2026-04-07 FirewallAnalyzer audit: `DriftDetector.RulesEqual` is a 14-line `=>`-bodied method with cyclomatic complexity 14 — exactly the kind of method users would want to flow-analyze). Add an internal lift step that wraps `=> expr` as `{ return expr; }` and runs flow analysis on the synthetic block, so callers can audit `=>`-bodied methods without rewriting the source. |
