+++ src/RoslynMcp.Roslyn/Helpers/NuGetVulnerabilityJsonParser.cs
@@ -1,0 +1,118 @@
+using System.Text.Json;
+using RoslynMcp.Core.Models;
+
+namespace RoslynMcp.Roslyn.Helpers;
+
+/// <summary>
+/// Parses the JSON payload returned by <c>dotnet list package --vulnerable --format json</c>
+/// into a deduplicated list of <see cref="NuGetVulnerablePackageDto"/> entries.
+/// Extracted from <c>DependencyAnalysisService</c> as part of the SRP split so the
+/// per-framework / per-package walk has its own testable surface and the cyclomatic
+/// complexity (~27 in the inline + nested-function form) drops to ~8.
+/// </summary>
+public static class NuGetVulnerabilityJsonParser
+{
+    /// <summary>
+    /// Parses the dotnet CLI JSON payload. <paramref name="scannedProjects"/> is set to the
+    /// number of projects in the payload's <c>projects</c> array.
+    /// </summary>
+    public static IReadOnlyList<NuGetVulnerablePackageDto> Parse(string json, out int scannedProjects)
+    {
+        scannedProjects = 0;
+        var results = new List<NuGetVulnerablePackageDto>();
+        var seen = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
+
+        using var doc = JsonDocument.Parse(json);
+        var root = doc.RootElement;
+        if (!root.TryGetProperty("projects", out var projectsEl) || projectsEl.ValueKind != JsonValueKind.Array)
+        {
+            return results;
+        }
+
+        scannedProjects = projectsEl.GetArrayLength();
+
+        foreach (var project in projectsEl.EnumerateArray())
+        {
+            var projectPath = project.TryGetProperty("path", out var pathProp) ? pathProp.GetString() ?? "" : "";
+            var projectName = string.IsNullOrEmpty(projectPath)
+                ? "unknown"
+                : Path.GetFileNameWithoutExtension(projectPath);
+
+            if (!project.TryGetProperty("frameworks", out var frameworks) || frameworks.ValueKind != JsonValueKind.Array)
+            {
+                continue;
+            }
+
+            foreach (var framework in frameworks.EnumerateArray())
+            {
+                AddPackagesFromFramework(framework, "topLevelPackages", isTransitive: false, projectPath, projectName, seen, results);
+                AddPackagesFromFramework(framework, "transitivePackages", isTransitive: true, projectPath, projectName, seen, results);
+            }
+        }
+
+        return results;
+    }
+
+    /// <summary>
+    /// Extracts vulnerable packages from one framework slice and appends new entries to
+    /// <paramref name="results"/>, keyed by (projectPath, id, resolved, advisoryUrl, severity, isTransitive)
+    /// to suppress duplicate rows that the dotnet CLI emits when a package is referenced
+    /// from more than one TFM in the same project.
+    /// </summary>
+    private static void AddPackagesFromFramework(
+        JsonElement framework,
+        string arrayName,
+        bool isTransitive,
+        string projectPath,
+        string projectName,
+        HashSet<string> seen,
+        List<NuGetVulnerablePackageDto> results)
+    {
+        if (!framework.TryGetProperty(arrayName, out var pkgs) || pkgs.ValueKind != JsonValueKind.Array)
+        {
+            return;
+        }
+
+        foreach (var pkg in pkgs.EnumerateArray())
+        {
+            var id = pkg.TryGetProperty("id", out var idEl) ? idEl.GetString() ?? "" : "";
+            var resolved = pkg.TryGetProperty("resolvedVersion", out var rv)
+                ? rv.GetString() ?? ""
+                : pkg.TryGetProperty("requestedVersion", out var rq)
+                    ? rq.GetString() ?? ""
+                    : "";
+
+            if (!pkg.TryGetProperty("vulnerabilities", out var vulns) || vulns.ValueKind != JsonValueKind.Array)
+            {
+                continue;
+            }
+
+            foreach (var vuln in vulns.EnumerateArray())
+            {
+                var severity = vuln.TryGetProperty("severity", out var sev)
+                    ? sev.GetString() ?? "Unknown"
+                    : "Unknown";
+                var advisoryUrl = vuln.TryGetProperty("advisoryurl", out var adv)
+                    ? adv.GetString()
+                    : vuln.TryGetProperty("advisoryUrl", out var adv2)
+                        ? adv2.GetString()
+                        : null;
+
+                var dedupeKey = $"{projectPath}|{id}|{resolved}|{advisoryUrl}|{severity}|{isTransitive}";
+                if (!seen.Add(dedupeKey))
+                {
+                    continue;
+                }
+
+                results.Add(new NuGetVulnerablePackageDto(
+                    id,
+                    resolved,
+                    severity,
+                    advisoryUrl,
+                    projectName,
+                    isTransitive,
+                    PatchedVersion: null));
+            }
+        }
+    }
+}
