+++ tests/RoslynMcp.Tests/NuGetVulnerabilityScanIntegrationTests.cs
@@ -18,7 +18,7 @@ public sealed class NuGetVulnerabilityScanIntegrationTests : SharedWorkspaceTest
     [TestMethod]
     public async Task ScanNuGetVulnerabilitiesAsync_ReturnsStructuredResult()
     {
-        var result = await DependencyAnalysisService.ScanNuGetVulnerabilitiesAsync(
+        var result = await NuGetDependencyService.ScanNuGetVulnerabilitiesAsync(
             WorkspaceId, projectFilter: null, includeTransitive: false, CancellationToken.None);
 
         Assert.IsNotNull(result);
